fix(deepseek): apply V4 HTTP compat on the regular router too
Move the V4 translation out of PDRouter into a shared
http/deepseek_compat module and call it from the regular Router's
send_typed_request for /v1/chat/completions, so non-PD HTTP workers
also receive the thinking kwarg projection and effort remap.
