+++ model_gateway/src/routers/http/mod.rs
@@ -1,5 +1,6 @@
 //! HTTP router implementations
 
+pub(crate) mod deepseek_compat;
 pub mod pd_router;
 pub mod pd_types;
 pub mod router;
